+++ src/main/java/epicsquid/mysticallib/particle/particles/ParticleGlow.java
@@ -10,7 +10,7 @@ public class ParticleGlow extends ParticleBase {
   public float initScale = 0;
   public float initAlpha = 0;
 
-  public ParticleGlow(World world, double x, double y, double z, double vx, double vy, double vz, double[] data) {
+  public ParticleGlow(World world, double x, double y, double z, double vx, double vy, double vz, float[] data) {
     super(world, x, y, z, vx, vy, vz, data);
     this.colorR = (float) data[1];
     this.colorG = (float) data[2];
